feat: add dispute resolution contract and update build configurations
